refactor(earn): co-locate lending testIDs for MMQA-1390
Move Earn lending/footer/progress/withdraw testIDs into *.testIds.ts
and point EarnLendingView PO at app imports; drop dead deposit-button.
