+++ app/components/UI/Earn/Views/EarnLendingDepositConfirmationView/components/ConfirmationFooter/ConfirmationFooter.testIds.ts
@@ -1,0 +1,11 @@
+export const CONFIRMATION_FOOTER_TEST_ID = 'earn-lending-confirmation-footer';
+
+export const CONFIRMATION_FOOTER_BUTTON_TEST_IDS = {
+  CANCEL_BUTTON: `${CONFIRMATION_FOOTER_TEST_ID}-cancel-button`,
+  CONFIRM_BUTTON: `${CONFIRMATION_FOOTER_TEST_ID}-confirm-button`,
+} as const;
+
+export const CONFIRMATION_FOOTER_LINK_TEST_IDS = {
+  TERMS_OF_USE_BUTTON: `${CONFIRMATION_FOOTER_TEST_ID}-terms-of-use-button`,
+  RISK_DISCLOSURE_BUTTON: `${CONFIRMATION_FOOTER_TEST_ID}-risk-disclosure-button`,
+} as const;
