+++ app/components/UI/Earn/Views/EarnLendingDepositConfirmationView/components/DepositInfoSection/DepositInfoSection.testIds.ts
@@ -1,0 +1,1 @@
+export const DEPOSIT_DETAILS_SECTION_TEST_ID = 'depositDetailsSection';
