+++ app/components/UI/Earn/Views/EarnLendingDepositConfirmationView/components/DepositReceiveSection/DepositReceiveSection.testIds.ts
@@ -1,0 +1,1 @@
+export const DEPOSIT_RECEIVE_SECTION_TEST_ID = 'depositReceiveSection';
