+++ app/components/UI/Earn/Views/EarnWithdrawInputView/EarnWithdrawInputView.testIds.ts
@@ -1,0 +1,3 @@
+export const EarnWithdrawInputViewTestIds = {
+  REVIEW_BUTTON: 'review-button',
+} as const;
