+++ app/components/UI/Earn/components/EarnLendingBalance/EarnLendingBalance.testIds.ts
@@ -1,0 +1,3 @@
+export const EARN_LENDING_BALANCE_TEST_IDS = {
+  WITHDRAW_BUTTON: 'withdraw-button',
+} as const;
